+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0 (01-05-2023)
+
+-   UI5 Parser updated to v1.1.0
+-   `Duplicate Translation Linter` added to properties linters
+-   Global config handler added
+
 ## 1.0.0 (13-04-2023)
 
 -   UI5 Parser updated to v1.0.0